// src/common/types.ts
export type Location = {
  pathname: string;
  search: string;
  hash: string;
};

export type BrowserHistory = (
  | Omit<History, 'location' | 'go' | 'createHref' | 'push' | 'replace'>
  | Omit<History5, 'location' | 'go' | 'createHref' | 'push' | 'replace'>
) & {
  location: Location;
  push: (path: string) => void;
  replace: (path: string) => void;
};

export type MatchParams = {
  [key: string]: string | null | typeof undefined;
};

export type Match = {
  /** TODO we are supporting `undefined` here because we are currently using both
   * this version of the `Match` type, and react-routers version (which allows for `undefined`)
   * To fix this we should move `matchPath` to our own util so we can apply our own types, then
   * decide if we want to support undefined types.
   */
  params: MatchParams;
  query: Query;
  isExact: boolean;
  path: string;
  url: string;
};

export type Query = {
  [key: string]: string;
};

export type MatchedInvariantRoute = {
  route: InvariantRoute;
  match: Match;
};

export type MatchedRoute = {
  route: Route;
  match: Match;
};

export type RouteContext = {
  location: Location;
  query: Query;
  route: Route;
  match: Match;
  action: HistoryAction;
};
// ...
export type RouteResource<T extends unknown = unknown> = {
  type: ResourceType;
  getKey: (
    routerContext: RouterContext,
    customContext: ResourceStoreContext
  ) => ResourceKey;
  maxAge: number;
  getData: (
    routerContext: RouterDataContext,
    customContext: ResourceStoreContext
  ) => T | Promise<T>;
  maxCache: number;
  isBrowserOnly: boolean;
  depends: ResourceType[] | null;
};

export type RouteResources = RouteResource[];
// ...
export type InvariantRoute = {
  path: string;
  exact?: boolean;

  /** Unique name for the route */
  name: string;

  /**
   * Query string matching. Each query param must match for the route to match.
   *
   * A query param can take the following shapes:
   *  * query name only: 'foo' - matches if query name 'foo' is present
   *  * query name matching value: 'foo=bar' - matches if query name 'foo' equals
   *    'bar' exactly
   *  * query name matching regex: 'foo=(bar.+) - matches if query name 'foo' equals
   *    regex '^(bar.+)$'. Note you must escape backslashes and wrap regex in parentheses.
   *  * query name NOT matching value: 'foo!=bar' - matches if query name 'foo' does
   *    not equal bar OR if query name 'foo' does not exist at all
   */
  query?: string[];
};

/**
 * Custom additional attributes that may be present in each route definition.
 */
export interface AdditionalRouteAttributes {}

export type Route = InvariantRoute & {
  /** The component to render on match, typed explicitly */
  component: ComponentType<RouteContext>;

  /**
   * The resources for the route
   */
  resources?: RouteResources;
} & AdditionalRouteAttributes;

export type HistoryAction = 'PUSH' | 'REPLACE' | 'POP' | '';

export type InvariantRoutes = InvariantRoute[];
export type Routes = Route[];

export type LinkElementType = 'a' | 'button';

export type LinkProps = AnchorHTMLAttributes<HTMLAnchorElement> & {
  children: ReactNode;
  target?: '_blank' | '_self' | '_parent' | '_top';
  href?: string;
  to?: string | Route | Promise<{ default: Route } | Route>;
  replace?: boolean;
  type?: 'a' | 'button';
  onClick?: (e: MouseEvent | KeyboardEvent) => void;
  onMouseEnter?: (e: MouseEvent) => void;
  onMouseLeave?: (e: MouseEvent) => void;
  onPointerDown?: (e: PointerEvent) => void;
  onFocus?: (e: FocusEvent) => void;
  onBlur?: (e: FocusEvent) => void;
  params?: MatchParams;
  query?: Query;
  prefetch?: false | 'hover' | 'mount';
};